todo: put lib9.SpectrumData class into deflib1.py
